% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{pca}
\alias{pca}
\title{Perform Principal Component Analysis (PCA)}
\usage{
pca(X, k)
}
\arguments{
\item{X}{A numeric matrix data or a data frame representing the input data}

\item{k}{An integer specifying the number of principal components to retain}
}
\value{
A matrix containing the projected data onto the selected principal components
}
\description{
This function performs Principal Component Analysis (PCA) on the input data matrix X
and returns the projected data onto the selected principal components
}
\examples{
# Load the iris dataset

data(iris)

#Extract the numerical features from the iris dataset
X <- iris[, 1:4]

#Apply PCA
result <- pca(X, k = 2)

#Extract the transformed data
transformed_data <- result$transformed_data

# Plot the transformed data
plot(transformed_data, col = iris$Species, pch = 19, xlab = "PC1", ylab = "PC2")

# Add legend
legend("topright", legend = levels(iris$Species), col = 1:3, pch = 19)

# Plot the eigenvectors
arrows(0, 0, result$eigenvectors[1, ], result$eigenvectors[2, ], length = 0.1, col = "blue")

}
